fix:[김민홍] 리뷰등록 기능 및 에러 수정[HEALTH-FOOD-PROCESS-20]

+++ src/main/java/com/example/sbsj_process/product/review/entity/ProductReview.java
@@ -38,7 +38,7 @@ public class ProductReview {
     private String context;
 
     @Column(length = 8, nullable = false)
-    private Long starRate;
+    private Double starRate;
 
     @Column(length = 8)
     private Long likeRecommend;
